Deadlock when LEAF has response permissions
The test is likely wrong but uncovers a deadlock. The leaf
has allow_response permissions. When a client sends back a reply
that is destined to a LEAF connection, since the LEAF has permissions
and is not allowed to publish on the reply subject, but it has
response permissions, the code tries to get a lock that is already
held which causes deadlock.

As I said, the test is probably improper because the replies are
tracked on the user connection, so even if we were to solve the
locking issue for the LEAF connection, the reply subject would
never be found and the publish always disallowed.

Signed-off-by: Ivan Kozlovic <ivan@synadia.com>